pyproject.toml
[tool.black]
line-length = 80
skip-magic-trailing-comma = true
skip-string-normalization = true
target-version = ['py310']

[tool.isort]
profile = "black"

[tool.poetry]
name = "feynmanium"
version = "0.1.0"
description = "A discord bot that works."
authors = ["Tony Brown <tb148@tuta.io>"]
license = "AGPL-3.0-or-later"
readme = "README.rst"
repository = "https://github.com/tb148/feynmanium"
include = ["README.rst"]

[tool.poetry.dependencies]
python = "^3.10"
toml = "^0.10.2"
sympy = "^1.10.1"
googletrans-py = "^4.0.0"
mpmath = "^1.2.1"
chess = "^1.9.2"
CairoSVG = "^2.5.2"
uvloop = "^0.16.0"
"discord.py" = "^2.0.0"
tomlkit = "^0.11.4"

[tool.poetry.dev-dependencies]
poetry = "^1.1.13"
black = ">=22.3.0"
isort = { extras = [
    "requirements_deprecated_finder",
    "pipfile_deprecated_finder"
], version = "^5.10.1" }
types-toml = "^0.10.8"
pylama = { extras = ["all"], version = "^8.4.1" }
pre-commit = "^2.20.0"

[tool.poetry.scripts]
feynmanium = "feynmanium.run:main"

[tool.pylama]
format = "pydocstyle"
linters = "eradicate,mccabe,pycodestyle,pydocstyle,pyflakes,pylint,isort"
max_line_length = 80

[tool.pylama.linter.pydocstyle]
convention = "google"

[tool.pylint.format]
max-line-length = 80

[tool.pyright]
typeCheckingMode = "basic"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
